[REFACTORING] Backend
