+++ backend/controllers/feedController.js
@@ -4,159 +4,21 @@ const path = require("path");
 const { Op } = require("sequelize");
 const {
   User,
-  ActivityLog,
   Quiz,
   UserLesson,
   UserAnswer,
   Question,
   Follow,
 } = require("../models");
+
 const { sendEmailConfirmation } = require("./userAuthController");
 
-// Functions
-
-const getUser = async (targetUserId) => {
-  const { id, first_name, last_name, email, avatar_url } = await User.findByPk(
-    targetUserId,
-    { paranoid: false }
-  );
-  return { id, first_name, last_name, email, avatar_url };
-};
-
-const getActivities = async (currentUserId, targetUserId) => {
-  try {
-    let result = [];
-    const user = await User.findByPk(targetUserId, { include: [ActivityLog] });
-    if (!user.ActivityLogs.length > 0) {
-      result = "No Activities yet";
-    }
-    await Promise.all(
-      user.ActivityLogs.map(async (activity) => {
-        if (activity.relatable_type === "quizzes") {
-          await Quiz.findByPk(activity.relatable_id, { paranoid: false }).then(
-            (quiz) => {
-              if (currentUserId === targetUserId) {
-                result = [
-                  ...result,
-                  {
-                    activity: `You learnt ${quiz.name}`,
-                    avatar_url: user.avatar_url,
-                    timestamp: activity.updatedAt,
-                  },
-                ];
-              } else {
-                result = [
-                  ...result,
-                  {
-                    activity: `${user.first_name} learnt ${quiz.name}`,
-                    avatar_url: user.avatar_url,
-                    timestamp: activity.updatedAt,
-                  },
-                ];
-              }
-            }
-          );
-        } else if (activity.relatable_type === "follows") {
-          const id = await Follow.findByPk(activity.relatable_id, {
-            attributes: ["following_id"],
-          });
-          await getUser(id.following_id).then(async (uInfo) => {
-            if (currentUserId === targetUserId) {
-              result = [
-                ...result,
-                {
-                  activity: `You followed ${uInfo.first_name}`,
-                  avatar_url: user.avatar_url,
-                  timestamp: activity.updatedAt,
-                },
-              ];
-            } else {
-              result = [
-                ...result,
-                {
-                  activity: `${user.first_name} followed ${uInfo.first_name}`,
-                  avatar_url: user.avatar_url,
-                  timestamp: activity.updatedAt,
-                },
-              ];
-            }
-          });
-        } else if (activity.relatable_type === "unfollows") {
-          const id = await Follow.findByPk(activity.relatable_id, {
-            attributes: ["following_id"],
-          });
-          await getUser(id.following_id).then(async (uInfo) => {
-            if (currentUserId === targetUserId) {
-              result = [
-                ...result,
-                {
-                  activity: `You unfollowed ${uInfo.first_name}`,
-                  avatar_url: user.avatar_url,
-                  timestamp: activity.updatedAt,
-                },
-              ];
-            } else {
-              result = [
-                ...result,
-                {
-                  activity: `${user.first_name} unfollowed ${uInfo.first_name}`,
-                  avatar_url: user.avatar_url,
-                  timestamp: activity.updatedAt,
-                },
-              ];
-            }
-          });
-        }
-      })
-    );
-    return result;
-  } catch (err) {
-    throw new Error({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getLearntWordsAndLessons = async (targetUserId) => {
-  try {
-    let numberOfWords = 0;
-    const user = await User.findByPk(targetUserId, { include: [UserLesson] });
-    const numberOfLessons = user.UserLessons.length;
-
-    await Promise.all(
-      user.UserLessons.map(async (lesson) => {
-        await UserAnswer.findAll({
-          where: { quiz_id: lesson.quiz_id, correct: true },
-        }).then((answers) => {
-          numberOfWords += answers.length;
-        });
-      })
-    );
-    return numberOfWords > 0
-      ? {
-          learntWordsResult: `Learnt ${numberOfWords} words`,
-          learntLessonsResult: `Learnt ${numberOfLessons} lessons`,
-        }
-      : "No Learnings yet";
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getFollowingIdLIst = async (user_id) => {
-  const followingIdList = [];
-  const following = await Follow.findAll({
-    where: { follower_id: user_id, flag: true },
-    attributes: ["following_id"],
-  });
-
-  if (following.length > 0) {
-    await following.map((m) => {
-      followingIdList.push(m.following_id);
-    });
-  }
-  return followingIdList;
-};
+const {
+  getActivities,
+  getLearntWordsAndLessons,
+  getFollowingIdLIst,
+  getUser,
+} = require("../utils/feed.utils");
 
 // Controllers
 
@@ -269,193 +131,6 @@ const getAllUsersInfo = async (req, res, next) => {
   }
 };
 
-const putfollowAndUnfollow = async (req, res, next) => {
-  const { id } = req.params;
-  const user_id = req.user;
-  try {
-    const check = await Follow.findOne({
-      where: { follower_id: user_id, following_id: id },
-      order: [["createdAt", "DESC"]],
-    });
-
-    if (check && check.flag) {
-      await Follow.update({ flag: false }, { where: { id: check.id } });
-      await ActivityLog.create({
-        relatable_id: check.id,
-        relatable_type: "unfollows",
-        user_id,
-      });
-      res.status(200).json({ message: "Unfollow operation succeded" });
-    } else if (check && check.flag === false) {
-      await Follow.update({ flag: true }, { where: { id: check.id } });
-      await ActivityLog.create({
-        relatable_id: check.id,
-        relatable_type: "follows",
-        user_id,
-      });
-      res.status(200).json({ message: "Follow operation succeded" });
-    } else {
-      const follow = await Follow.create({
-        follower_id: user_id,
-        following_id: id,
-        flag: true,
-      });
-      await ActivityLog.create({
-        relatable_id: follow.id,
-        relatable_type: "follows",
-        user_id: user_id,
-      });
-      res.status(200).json({ message: "Follow operation succeded" });
-    }
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getFollowsCount = async (req, res, next) => {
-  const { id } = req.params;
-  try {
-    const following = await Follow.findAll({
-      where: { follower_id: id, flag: true },
-    });
-    let followingCount = 0;
-    await Promise.all(
-      following.map(async ({ following_id }) => {
-        const user = await User.findByPk(following_id);
-        if (user) {
-          followingCount++;
-        }
-      })
-    );
-
-    const follower = await Follow.findAll({
-      where: { following_id: id, flag: true },
-    });
-    let followerCount = 0;
-    await Promise.all(
-      follower.map(async ({ follower_id }) => {
-        const user = await User.findByPk(follower_id);
-        if (user) {
-          followerCount++;
-        }
-      })
-    );
-
-    res.status(200).json({
-      followers: followerCount,
-      following: followingCount,
-    });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getFollowing = async (req, res, next) => {
-  const { id } = req.params;
-  const user_id = req.user;
-  const followingsList = [];
-
-  try {
-    const followingIdList = await getFollowingIdLIst(user_id);
-
-    const { Follows } = await User.findByPk(id, {
-      include: [{ model: Follow, attributes: ["following_id", "flag"] }],
-    });
-
-    if (Follows.length > 0) {
-      await Promise.all(
-        Follows.map(async (follow) => {
-          if (follow.flag) {
-            const { id, first_name, last_name, email, avatar_url, deletedAt } =
-              await User.findByPk(follow.following_id, { paranoid: false });
-            if (deletedAt) {
-              return;
-            }
-            if (followingIdList.includes(id)) {
-              followingsList.push({
-                id,
-                first_name,
-                last_name,
-                email,
-                avatar_url,
-                follows: true,
-              });
-            } else {
-              followingsList.push({
-                id,
-                first_name,
-                last_name,
-                email,
-                avatar_url,
-                follows: false,
-              });
-            }
-          }
-        })
-      );
-    }
-
-    res.status(200).json({ following: followingsList });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getFollowers = async (req, res, next) => {
-  const { id } = req.params;
-  const user_id = req.user;
-  const followersList = [];
-
-  try {
-    const followingIdList = await getFollowingIdLIst(user_id);
-
-    const follower = await Follow.findAll({
-      where: { following_id: id, flag: true },
-      attributes: ["follower_id"],
-    });
-
-    if (follower.length > 0) {
-      await Promise.all(
-        follower.map(async (m) => {
-          const { id, first_name, last_name, email, avatar_url } =
-            await User.findByPk(m.follower_id);
-          if (followingIdList.includes(id)) {
-            followersList.push({
-              id,
-              first_name,
-              last_name,
-              email,
-              avatar_url,
-              follows: true,
-            });
-          } else {
-            followersList.push({
-              id,
-              first_name,
-              last_name,
-              email,
-              avatar_url,
-              follows: false,
-            });
-          }
-        })
-      );
-    }
-
-    res.status(200).json({ followers: followersList });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
 const editPersonalInfo = async (req, res, next) => {
   const user_id = req.user;
   const first_name = req.body.first_name;
@@ -553,370 +228,6 @@ const editPassword = async (req, res, next) => {
   }
 };
 
-const getCategories = async (req, res, next) => {
-  const { search, orderBy } = req.query;
-  const user_id = req.user;
-
-  const searchQuery = {
-    [Op.or]: {
-      name: { [Op.like]: `%${search}%` },
-      description: { [Op.like]: `%${search}%` },
-    },
-  };
-
-  const orderByQuery = [["name", orderBy]];
-
-  try {
-    const lessons = await UserLesson.findAll({
-      where: { user_id },
-      attributes: ["quiz_id"],
-    });
-
-    const answers = await UserAnswer.findAll({
-      where: { user_id, correct: true },
-      include: [Question],
-    });
-
-    const takenLessonsList = [];
-    lessons.map((lesson) => {
-      takenLessonsList.push(lesson.quiz_id);
-    });
-
-    const categories = await Quiz.findAll({
-      include: [Question],
-      where: search ? searchQuery : { name: { [Op.ne]: null } },
-      order: orderBy ? orderByQuery : [["id", "ASC"]],
-    });
-    const result = [];
-    await Promise.all(
-      categories.map(async ({ id, name, description, Questions }) => {
-        if (takenLessonsList.includes(id)) {
-          let score = 0;
-          await Promise.all(
-            answers
-              .filter((answer) => answer.Question !== null)
-              .map(async (answer) => {
-                const correct_answer = await Question.findOne({
-                  where: { id: answer.question_id },
-                  attributes: [answer.Question.correct_answer],
-                });
-                if (
-                  answer.quiz_id === id &&
-                  answer.user_answer ===
-                    correct_answer[answer.Question.correct_answer]
-                ) {
-                  score++;
-                }
-              })
-          );
-          result.push({
-            id,
-            name,
-            description,
-            wasTaken: true,
-            score,
-            questionsCount: Questions.length,
-          });
-        } else {
-          result.push({
-            id,
-            name,
-            description,
-            wasTaken: false,
-            score: null,
-            questionsCount: Questions.length,
-          });
-        }
-      })
-    );
-
-    res.status(200).json({ categories: result });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getLesson = async (req, res, next) => {
-  const { id } = req.params;
-  const user_id = req.user;
-
-  try {
-    const userlesson = await UserLesson.findOne({
-      where: { quiz_id: id, user_id },
-    });
-    const userAnwers = await UserAnswer.findAll({
-      where: { quiz_id: id, user_id },
-    });
-
-    if (userAnwers.length > 0) {
-      await UserAnswer.destroy({
-        where: { quiz_id: id, user_id },
-      });
-    }
-
-    if (!userlesson) {
-      await UserLesson.create({
-        user_id,
-        quiz_id: id,
-        score: 0,
-      });
-    } else {
-      await userlesson.update({
-        score: 0,
-      });
-    }
-
-    await ActivityLog.create({
-      relatable_id: id,
-      relatable_type: "quizzes",
-      user_id,
-    });
-
-    const { name, Questions } = await Quiz.findByPk(id, {
-      include: [Question],
-      attributes: ["name"],
-    });
-
-    const result = [];
-
-    Questions.map(({ id, title, choice_1, choice_2, choice_3, choice_4 }) => {
-      result.push({
-        id,
-        title,
-        choices: [choice_1, choice_2, choice_3, choice_4],
-      });
-    });
-
-    res.status(200).json({ questions: result, name });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getQuestions = async (req, res, next) => {
-  const { id } = req.params;
-
-  try {
-    const { name, Questions } = await Quiz.findByPk(id, {
-      include: [Question],
-      attributes: ["name"],
-    });
-
-    const result = [];
-
-    Questions.map(
-      ({
-        id,
-        title,
-        choice_1,
-        choice_2,
-        choice_3,
-        choice_4,
-        correct_answer,
-      }) => {
-        result.push({
-          id,
-          title,
-          choices: [choice_1, choice_2, choice_3, choice_4],
-          correct_answer,
-        });
-      }
-    );
-
-    res.status(200).json({ questions: result, name });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const postAnswer = async (req, res, next) => {
-  const user_id = req.user;
-  const { question_id, user_answer, quiz_id } = req.body;
-
-  try {
-    const answer = await UserAnswer.findOne({
-      where: { user_id, question_id, quiz_id },
-    });
-
-    const question = await Question.findOne({
-      where: { id: question_id, quiz_id },
-    });
-
-    const correct_answer = await Question.findOne({
-      where: { id: question_id },
-      attributes: [question.correct_answer],
-    });
-
-    if (answer) {
-      if (answer.user_answer !== user_answer) {
-        if (user_answer === correct_answer[question.correct_answer]) {
-          await answer.update({
-            user_answer,
-            correct: true,
-          });
-        } else {
-          await answer.update({
-            user_answer,
-            correct: false,
-          });
-        }
-
-        return res.status(200).json({ message: "Answer updated!" });
-      } else {
-        return res.status(304).json({ error: "Answer exists!" });
-      }
-    } else {
-      if (user_answer === correct_answer[question.correct_answer]) {
-        await UserAnswer.create({
-          user_id,
-          question_id,
-          user_answer,
-          quiz_id,
-          correct: true,
-        });
-      } else {
-        await UserAnswer.create({
-          user_id,
-          question_id,
-          user_answer,
-          quiz_id,
-          correct: false,
-        });
-      }
-
-      res.status(200).json({ message: "Answer received!" });
-    }
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getAnswer = async (req, res, next) => {
-  const user_id = req.user;
-  const { quiz_id } = req.query;
-
-  try {
-    const answers = await UserAnswer.findAll({
-      where: { user_id, quiz_id },
-      attributes: ["user_answer", "question_id", "quiz_id"],
-    });
-
-    res.status(200).json({ answers });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
-const getResult = async (req, res, next) => {
-  const user_id = req.user;
-  const { quiz_id } = req.params;
-
-  try {
-    const userlesson = await UserLesson.findOne({
-      where: { quiz_id, user_id },
-    });
-
-    if (!userlesson) {
-      return res.status(404).json({ error: "Quiz was not taken" });
-    }
-
-    const quiz = await Quiz.findByPk(quiz_id, {
-      include: [
-        { model: UserAnswer, where: { user_id } },
-        { model: Question, where: { quiz_id } },
-      ],
-    });
-
-    const result = [];
-    let score = 0;
-
-    if (userlesson && !quiz) {
-      const questions = await Question.findAll({ where: { quiz_id } });
-      await Promise.all(
-        questions.map(async (question) => {
-          const correct_answer = await Question.findOne({
-            where: { id: question.id },
-            attributes: [question.correct_answer],
-          });
-
-          result.push({
-            question_id: question.id,
-            correct: false,
-            correct_answer: correct_answer[question.correct_answer],
-            user_answer: null,
-          });
-        })
-      );
-      return res.status(200).json({ result, score });
-    }
-
-    const answeredQuestionIdList = [];
-    quiz.UserAnswers.map((user_answer) => {
-      answeredQuestionIdList.push(user_answer.question_id);
-    });
-    await Promise.all(
-      quiz.Questions.map(async (question) => {
-        const correct_answer = await Question.findOne({
-          where: { id: question.id },
-          attributes: [question.correct_answer],
-        });
-        if (!answeredQuestionIdList.includes(question.id)) {
-          result.push({
-            question_id: question.id,
-            correct: false,
-            correct_answer: correct_answer[question.correct_answer],
-            user_answer: null,
-          });
-        }
-        await quiz.UserAnswers.map((answer) => {
-          if (
-            answer.question_id === question.id &&
-            answer.user_answer === correct_answer[question.correct_answer]
-          ) {
-            score += 1;
-            result.push({
-              question_id: question.id,
-              correct: true,
-              correct_answer: correct_answer[question.correct_answer],
-              user_answer: answer.user_answer,
-            });
-          } else if (
-            answer.question_id === question.id &&
-            answer.user_answer !== correct_answer[question.correct_answer]
-          ) {
-            result.push({
-              question_id: question.id,
-              correct: false,
-              correct_answer: correct_answer[question.correct_answer],
-              user_answer: answer.user_answer,
-            });
-          }
-        });
-      })
-    );
-
-    if (userlesson !== score) {
-      await userlesson.update({ score });
-    }
-
-    res.status(200).json({ result, score });
-  } catch (err) {
-    res
-      .status(400)
-      .json({ error: "Something went wrong, please try again later" });
-  }
-};
-
 const getLearntWords = async (req, res, next) => {
   const { user_id } = req.params;
 
@@ -998,18 +309,8 @@ module.exports = {
   getLearnigsCount,
   getUserInfo,
   getAllUsersInfo,
-  putfollowAndUnfollow,
-  getFollowsCount,
-  getFollowing,
-  getFollowers,
   editPersonalInfo,
   editEmail,
   editPassword,
-  getCategories,
-  getLesson,
-  postAnswer,
-  getAnswer,
-  getResult,
   getLearntWords,
-  getQuestions,
 };
