+++ frontend/src/pages/result/Result.js
@@ -25,15 +25,21 @@ const Result = () => {
           )}
 
           {resultData && data
-            ? resultData.result.map((result, i) => (
-                <BodySection
-                  key={result.question_id}
-                  question={data.questions[i]}
-                  user_answer={result.user_answer}
-                  correct_answer={result.correct_answer}
-                  correct={result.correct}
-                />
-              ))
+            ? resultData.result
+                .sort((a, b) => a.question_id - b.question_id)
+                .map((result, i) => (
+                  <BodySection
+                    key={result.question_id}
+                    question={
+                      data.questions.filter(
+                        (question) => question.id === result.question_id
+                      )[0]
+                    }
+                    user_answer={result.user_answer}
+                    correct_answer={result.correct_answer}
+                    correct={result.correct}
+                  />
+                ))
             : null}
           {isLoading || questionsLoading ? (
             <ArrowPathIcon className="w-8 animate-spin mx-auto my-2" />
